refactor: consolidate instance linking utilities
