+++ apps/builder/app/shared/instance-utils/link.ts
@@ -1,3 +1,9 @@
+import {
+  getAllPages,
+  isPageTemplate,
+  type Instances,
+  type Pages,
+} from "@webstudio-is/sdk";
 import type { InstanceSelector } from "@webstudio-is/project-build/runtime";
 import {
   $authToken,
@@ -6,7 +12,7 @@ import {
 } from "../nano-states";
 import { builderUrl } from "../router-utils";
 import { $instances, $pages, $project } from "../sync/data-stores";
-import { getDeepLinkedInstanceSelection } from "./selection";
+import { canResolveInstanceSelector } from "./selection";
 
 export const getInstanceLink = (
   instanceSelector: InstanceSelector | undefined
@@ -35,4 +41,44 @@ export const getInstanceLink = (
     safemode:
       new URLSearchParams(window.location.search).get("safemode") === "true",
   });
+};
+
+export const getInstanceSelectorFromUrl = (searchParams: URLSearchParams) =>
+  searchParams.get("instance")?.split(",");
+
+export const getDeepLinkedInstanceSelection = ({
+  instanceSelector,
+  canOpenPageTemplates,
+  pages,
+  instances,
+}: {
+  instanceSelector: InstanceSelector | undefined;
+  canOpenPageTemplates: boolean;
+  pages: Pages;
+  instances: Instances;
+}) => {
+  if (instanceSelector === undefined) {
+    return;
+  }
+
+  const instanceId = instanceSelector[0];
+  if (
+    instanceId === undefined ||
+    instances.has(instanceId) === false ||
+    canResolveInstanceSelector(instanceSelector, instances) === false
+  ) {
+    return;
+  }
+
+  const rootInstanceId = instanceSelector.at(-1);
+  const page = getAllPages(pages).find(
+    (page) => page.rootInstanceId === rootInstanceId
+  );
+  if (
+    page === undefined ||
+    (isPageTemplate(page) && canOpenPageTemplates === false)
+  ) {
+    return;
+  }
+  return { pageId: page.id, instanceSelector };
 };
